# src/docs/swagger.yaml
definitions:
  forms.Result:
    properties:
      error:
        example: some error note
        type: string
      result:
        example: some success note
        type: string
      status:
        example: OK
        type: string
    type: object
  forms.User:
    properties:
      email:
        example: hello@example.com
        type: string
      id:
        example: 5125112
        type: integer
      password:
        example: 1fsgh2rfafas
        type: string
      username:
        example: Martin
        type: string
    type: object
  models.Album:
    properties:
      Date:
        example: 0
        type: integer
      authorId:
        example: 121
        type: integer
      countLikes:
        example: 54
        type: integer
      countListening:
        example: 15632
        type: integer
      coverId:
        example: 254
        type: integer
      id:
        example: 777
        type: integer
      title:
        example: Mercury
        type: string
    type: object
...
  utils.Error:
    properties:
      err:
        type: string
    type: object
  utils.Success:
    properties:
      result: {}
    type: object
info:
  contact: {}
paths:
  /net/v1/albums/:
    get:
      consumes:
        - application/json
      description: getting all albums
      produces:
        - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/utils.Success'
        "400":
          description: Data is invalid
          schema:
            $ref: '#/definitions/utils.Error'
        "405":
          description: Method is not allowed
          schema:
            $ref: '#/definitions/utils.Error'
      summary: GetAlbums
      tags:
      - album
    post:
      consumes:
      - application/json
      description: creating new album
      parameters:
      - description: params of new album. Id will be set automatically.
        in: body
        name: Album
        required: true
        schema:
          $ref: '#/definitions/models.Album'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/utils.Success'
        "400":
          description: Data is invalid
          schema:
            $ref: '#/definitions/utils.Error'
        "405":
          description: Method is not allowed
          schema:
            $ref: '#/definitions/utils.Error'
      summary: CreateAlbum
      tags:
      - album
    put:
      consumes:
      - application/json
      description: updating album by id
      parameters:
      - description: id of updating album and params of it.
        in: body
        name: Album
        required: true
        schema:
          $ref: '#/definitions/models.Album'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/utils.Success'
        "400":
          description: Data is invalid
          schema:
            $ref: '#/definitions/utils.Error'
        "405":
          description: Method is not allowed
          schema:
            $ref: '#/definitions/utils.Error'
      summary: UpdateAlbum
      tags:
      - album
  /net/v1/albums/{id}:
    delete:
      consumes:
        - application/json
      description: deleting album by id
      parameters:
        - description: id of album which need to be deleted
          in: path
          name: id
          required: true
          type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/utils.Success'
        "400":
          description: Data is invalid
          schema:
            $ref: '#/definitions/utils.Error'
        "405":
          description: Method is not allowed
          schema:
            $ref: '#/definitions/utils.Error'
      summary: DeleteAlbum
      tags:
      - album
    get:
      consumes:
      - application/json
      description: getting album by id
      parameters:
      - description: id of album which need to be getted
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Album'
        "400":
          description: Data is invalid
          schema:
            $ref: '#/definitions/utils.Error'
        "405":
          description: Method is not allowed
          schema:
            $ref: '#/definitions/utils.Error'
      summary: GetAlbum
      tags:
      - album
  /net/v1/albums/popular:
    get:
      consumes:
        - application/json
      description: getting top20 popular albums
      produces:
        - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/utils.Success'
        "400":
          description: Data is invalid
          schema:
            $ref: '#/definitions/utils.Error'
        "405":
          description: Method is not allowed
          schema:
            $ref: '#/definitions/utils.Error'
      summary: GetPopularAlbums
      tags:
      - album
...
  /net/v1/login:
    post:
      consumes:
        - application/json
      description: login user
      parameters:
        - description: user params
          in: body
          name: UserForm
          required: true
          schema:
          $ref: '#/definitions/forms.User'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/forms.Result'
        "400":
          description: invalid login or password
          schema:
            $ref: '#/definitions/forms.Result'
        "401":
          description: invalid csrf
          schema:
            $ref: '#/definitions/forms.Result'
      summary: Login
      tags:
      - auth
  /net/v1/logout:
    post:
      consumes:
        - application/json
      description: sign in user
      produces:
        - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/forms.Result'
        "401":
          description: invalid csrf
          schema:
            $ref: '#/definitions/forms.Result'
        "460":
          description: Data is invalid
          schema:
            $ref: '#/definitions/forms.Result'
        "521":
          description: Cannot create session
          schema:
            $ref: '#/definitions/forms.Result'
      summary: Logout
      tags:
      - auth
  /net/v1/signup:
    post:
      consumes:
        - application/json
      description: login user
      parameters:
        - description: new jwt data
          in: body
          name: UserForm
          required: true
          schema:
          $ref: '#/definitions/forms.User'
      produces:
      - application/json
      responses:
        "200":
          description: success signup
          schema:
            $ref: '#/definitions/forms.Result'
        "400":
          description: user already exist
          schema:
            $ref: '#/definitions/forms.Result'
        "401":
          description: invalid csrf
          schema:
            $ref: '#/definitions/forms.Result'
      summary: SignUp
      tags:
      - auth
...
swagger: "2.0"